Reload page on version update
This is normally done by Next.js automatically, but is aparently
not working for statically exported pages, causing the site to
get stuck when trying to follow a link after the website has been
updated in the background.

https://github.com/zeit/next.js/issues/3558

As a workaround, we speculatively do a full reload five seconds after
a route change has started but not completed. To the user, this will
seem like just a particularly long page load.