Add script for importing the exported blog posts
This is not perfect by any means, but good enough that I already
run it on our Contentful space. The main remaining problem is
that many posts have internal links that look like this:

```
<link 57 - internal link "Open">Mehr Informationen...</link>
```

where 57 is the internal ID of the target on the old website...
currently, this script just links all of those to a non-existing
URL (/blog/internal/$ID).

The script is written to be idempotent (it replaces the
already-imported posts instead of creating a new set every run),
so we can always improve it and run the script again.