Switch over to new Contentful space with default locale set to German
This will let us finally have the blog posts fall back to the German versions when browsing the site in English. It also makes CMS editing a lot cleaner, because it will always be enough to just enter the German fields. The individual blog post fetches still have to have a manual fallback in the code though, since we're querying on slug, and when the slug just doesn't exist in English, the locale fallback mechanism doesn't kick in.