Remove withPreview check to fix broken blog reloads
The forced reload in componentDidMount that withPreviewCheck does is using a __NEXT_DATA__ variable, which only contains the right values for the initially loaded static page. As soon as we follow a link, __NEXT_DATA__ does not contain the corresponding URL query etc. Removing this means that Contentful previews will be damn slow again, but I see no other quick way of fixing this now, except teaching the Masifunde editors that they can quickly navigate away and back using a footer link to see the updated content.