Upgrade to Next.js 9.5
Next.js 9.3 introduced `getStaticProps` and `getStaticPaths`, which
will replace `getInitialProps` for static pages. Migrate all pages
to it.

Unfortunately, `exportPathMap` is not compatible with `getStaticPaths`.
This makes it impossible to map arbitrary paths to pages. For this
reason, all pages are now in `pages/[locale]`, and redirects are
used to map from `/german/path` to `/de/german/path`.

Also, update to styled-components v4, which made it easier to retain
support for server-side style generation.