Replace custom donation form with Fundraisingbox form
Fundraisingbox offers the "Polaris" theme now, which is mobile-friendly
and easy to style with CSS variables. For this reason, a custom form
for entering the donation and personal data is not necessary anymore.

By using the native Fundraisingbox form, it is possible to customize
the form directly from the Fundraisingbox configuration without any
development effort.